add tls support for *kellner*
using https to protect delivery of packages from certain attacks is best
practice. one way to achieve this is to put a https-enabled reverse proxy
in front of *kellner* and done. on the hand: in order to add (yet to come)
functionality to *kellner* itself, builtin support for tls is required and
that's what this commit does.