improve logging
to log all requests it's better to have the logger() wrapper as the final
wrapper around all other handlers. to move it into main() also reduces
code smell in the AttachHttpHandler() function.

this commit also adds an experimental 'feature' to logging: if a handler
sets the header-key "kellner-log-data" it will automatically pop up in
the logging output. data set to that key by the http-clients is discarded.