improve logging: add -log <filename> flag
in situations where *kellner* is running and the stderr/stdout are not
captured, logging to a file becomes quite useful. this commit adds the
-log flag.

sending SIGUSR1 closes the logfile, truncates it and recreates it. this
allows programs like logrotate to pick up the logfile, rename it and
keeping the logentries coming.