rewrite: use a cache-folder to store the extracted control-files
restarting *kellner* on a package folder with ~15k packages might take
too long (4 packages/s ~ 60 minutes). to improve the time to rescan
packages *kellner* uses a caching folder to store the extracted meta
information per package.

the Packages, Packages.gz and Packages.stamps files are not stored in
memory only but are created on the fly for each request based upon the
caching folder.

also -prep-cache can be used to prepare a caching folder.